+++ formula/src/main/java/com/instacart/formula/FormulaContext.kt
@@ -1,9 +1,9 @@
 package com.instacart.formula
 
-import com.instacart.formula.internal.JoinedKey
-import com.instacart.formula.internal.ScopedListeners
+import com.instacart.formula.internal.Listeners
 import com.instacart.formula.internal.TransitionDispatcher
 import com.instacart.formula.internal.UnitListener
+import kotlin.reflect.KClass
 
 /**
  * Provides functionality within [evaluate][Formula.evaluate] function to [compose][child]
@@ -11,7 +11,7 @@ import com.instacart.formula.internal.UnitListener
  * to arbitrary asynchronous events.
  */
 abstract class FormulaContext<out Input, State> internal constructor(
-    @PublishedApi internal val listeners: ScopedListeners,
+    @PublishedApi internal val listeners: Listeners,
     internal val transitionDispatcher: TransitionDispatcher<Input, State>,
 ) {
 
@@ -24,7 +24,7 @@ abstract class FormulaContext<out Input, State> internal constructor(
         transition: Transition<Input, State, Event>,
     ): Listener<Event> {
         return eventListener(
-            key = transition.type(),
+            key = ensureKeyIsScoped(transition.type()),
             transition = transition
         )
     }
@@ -39,7 +39,7 @@ abstract class FormulaContext<out Input, State> internal constructor(
         transition: Transition<Input, State, Event>,
     ): Listener<Event> {
         return eventListener(
-            key = JoinedKey(key, transition.type()),
+            key = ensureKeyIsScoped(transition.type(), key),
             transition = transition
         )
     }
@@ -129,9 +129,11 @@ abstract class FormulaContext<out Input, State> internal constructor(
      * @param key Unique identifier that will be used for this block.
      */
     inline fun <Value> key(key: Any, create: () -> Value): Value {
-        listeners.enterScope(key)
+        ensureNotRunning()
+
+        enterScope(key)
         val value = create()
-        listeners.endScope()
+        endScope()
         return value
     }
 
@@ -139,9 +141,18 @@ abstract class FormulaContext<out Input, State> internal constructor(
         key: Any,
         transition: Transition<Input, State, Event>
     ): Listener<Event> {
+        ensureNotRunning()
         val listener = listeners.initOrFindListener<Input, State, Event>(key)
         listener.transitionDispatcher = transitionDispatcher
         listener.transition = transition
         return listener
     }
+
+    // Internal key scope management
+    @PublishedApi internal abstract fun enterScope(key: Any)
+    @PublishedApi internal abstract fun endScope()
+    @PublishedApi internal abstract fun ensureKeyIsScoped(type: KClass<*>, key: Any? = null): Any
+
+    // Internal validation
+    @PublishedApi internal abstract fun ensureNotRunning()
 }