+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Add `action.runAgain()` extension function
 - Add `action.cancelPrevious(previousAction)` extension function
 - Add `DelegateAction` abstract class
+- Add child formula support for `context.key` scopes.
+- Add support for nested `context.key` calls.
 
 ## [0.7.0] - June 30, 2021
 - **Breaking**: Remove `events(observable) { }` utility function.